+++ src/components/homepage/SectionMusic.js
@@ -25,7 +25,7 @@ export function SectionMusic() {
       <div className="flex justify-center mt-10">
         <Link href="/music">
           <div>
-            <span className="text-md bg-neutral-800 text-white p-3 rounded-md hover:bg-neutral-600 duration-300">
+            <span className="text-md bg-neutral-800 text-white p-3 rounded-md hover:bg-neutral-800/80 duration-250">
               All Music
             </span>
           </div>